Dodana funkcionalnost da admin ne moze brisati sam sebe

+++ frontend/src/components/Admin.jsx
@@ -6,7 +6,7 @@ import { useAuth } from "../context/AuthContext"
 const Admin = () => {
   const [users, setUsers] = useState([]);
   const [query, setQuery] = useState("");
-  const { token } = useAuth();
+  const { token, email } = useAuth();
 
   useEffect(() => {
     if (token) {
@@ -22,8 +22,13 @@ const Admin = () => {
     }
   }, [token]);
 
-  const deleteUser = (email) => {
-    fetch(`/api/deleteUser/${email}`, {
+  const deleteUser = (userEmail) => {
+    if (userEmail === email) {
+      alert("Ne možete obrisati sami sebe!")
+      return;
+    }
+
+    fetch(`/api/deleteUser/${userEmail}`, {
       method: 'DELETE',
       headers: {
         Authorization: `Bearer ${token}`,
@@ -33,7 +38,7 @@ const Admin = () => {
           if (!response.ok) {
             throw new Error('Network response was not ok');
           }
-          setUsers(users.filter(user => user.email !== email));
+          setUsers(users.filter(user => user.email !== userEmail));
         })
         .catch((error) => console.error('Error:', error));
   };
